# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/catalog-policies/dry-run": {
      "post": {
        "operationId": "DryRunController_executeDryRun",
        "summary": "Execute dry-run evaluation",
        "description": "Evaluates media items against a proposed policy WITHOUT persisting results. Supports multiple selection modes: sample (random), top (by popularity), byType, byCountry. Limits: max 10000 items, 60s timeout.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Proposed policy and dry-run options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DryRunRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dry-run results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DryRunResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin - Policy Activation"
        ]
      }
    },
    "/api/admin/catalog-policies/dry-run/diff": {
      "post": {
        "operationId": "DryRunController_executeDryRunDiff",
        "summary": "Execute dry-run with diff",
        "description": "Same as dry-run but also includes comparison against current active policy version.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Proposed policy and dry-run options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DryRunRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dry-run results with current policy version",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DryRunResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin - Policy Activation"
        ]
      }
    },
    "/api/admin/catalog-policies/backfill/content-class": {
      "post": {
        "operationId": "BackfillController_backfillContentClass",
        "summary": "Backfill content_class for all media items",
        "description": "Classifies all existing media items based on genres and origin metadata. Use dry-run mode to preview changes without updating the database. Set triggerReEvaluation=true to automatically re-evaluate catalog after backfill.",
        "parameters": [
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "description": "If true, only simulates changes without updating DB",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "batchSize",
            "required": false,
            "in": "query",
            "description": "Batch size for processing (default: 1000)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "triggerReEvaluation",
            "required": false,
            "in": "query",
            "description": "If true, triggers catalog re-evaluation after backfill completes",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Backfill completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BackfillResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin - Content Classification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ratingo API",
    "description": "Rest API for Ratingo mobile and web clients",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MinVotesAnyOfDto": {
        "type": "object",
        "properties": {
          "sources": {
            "type": "array",
            "description": "Vote sources to check (OR logic)",
            "example": [
              "imdb",
              "trakt"
            ],
            "items": {
              "type": "string",
              "enum": [
                "imdb",
                "trakt"
              ]
            }
          },
          "min": {
            "type": "number",
            "description": "Minimum votes threshold",
            "example": 3000,
            "minimum": 0
          }
        }
      },
      "GlobalRequirementsDto": {
        "type": "object",
        "properties": {
          "minQualityScoreNormalized": {
            "type": "number",
            "description": "Minimum quality score normalized (0-1)",
            "example": 0.6,
            "minimum": 0,
            "maximum": 1
          },
          "requireAnyOfRatingsPresent": {
            "type": "array",
            "description": "At least one of these rating sources must be present (OR logic)",
            "example": [
              "imdb",
              "trakt"
            ],
            "items": {
              "type": "string",
              "enum": [
                "imdb",
                "metacritic",
                "rt",
                "trakt"
              ]
            }
          },
          "minVotesAnyOf": {
            "description": "Minimum votes from ANY of the specified sources (OR logic). Passes if any source meets the threshold. Robust to missing data.",
            "allOf": [
              {
                "$ref": "#/components/schemas/MinVotesAnyOfDto"
              }
            ]
          },
          "appliesTo": {
            "type": "array",
            "description": "Contexts where global gate applies. Defaults to ['catalog', 'homepage', 'trending', 'search'] if not specified. Freshness surfaces (now_playing, new_digital) are excluded by default.",
            "example": [
              "catalog",
              "homepage",
              "trending",
              "search"
            ],
            "items": {
              "type": "string",
              "enum": [
                "catalog",
                "homepage",
                "trending",
                "now_playing",
                "new_digital",
                "search"
              ]
            }
          }
        }
      },
...
      "CreatePolicyDto": {
        "type": "object",
        "properties": {
          "allowedCountries": {
            "description": "Allowed countries (ISO 3166-1 alpha-2 codes)",
            "example": [
              "US",
              "GB",
              "CA",
              "AU",
              "UA"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "blockedCountries": {
            "description": "Blocked countries (ISO 3166-1 alpha-2 codes)",
            "example": [
              "RU",
              "BY"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "blockedCountryMode": {
            "type": "string",
            "description": "Blocked country mode",
            "example": "ANY",
            "enum": [
              "ANY",
              "MAJORITY"
            ]
          },
          "allowedLanguages": {
            "description": "Allowed languages (ISO 639-1 codes)",
            "example": [
              "en",
              "uk",
              "de",
              "fr"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "blockedLanguages": {
            "description": "Blocked languages (ISO 639-1 codes)",
            "example": [
              "ru"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "globalProviders": {
            "description": "Global streaming providers",
            "example": [
              "netflix",
              "max",
              "appletv",
              "prime",
              "disney"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "breakoutRules": {
            "description": "Breakout rules for exceptions",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "eligibilityMode": {
            "type": "string",
            "description": "Eligibility mode",
            "example": "STRICT",
            "enum": [
              "STRICT",
              "RELAXED"
            ]
          },
          "homepage": {
            "type": "object",
            "description": "Homepage configuration",
            "example": {
              "minRelevanceScore": 50
            }
          },
          "globalRequirements": {
            "description": "Global quality gate requirements (all conditions combined with AND)",
            "allOf": [
              {
                "$ref": "#/components/schemas/GlobalRequirementsDto"
              }
            ]
          },
          "excludedContentClasses": {
            "type": "array",
            "description": "Content classes to exclude from catalog. SOFT filter: breakout rules CAN override.",
            "example": [
              "anime",
              "reality"
            ],
            "items": {
              "type": "string",
              "enum": [
                "mainstream",
                "anime",
                "documentary",
                "reality",
                "kids"
              ]
            }
          }
        },
        "required": [
          "allowedCountries",
          "blockedCountries",
          "allowedLanguages",
          "blockedLanguages"
        ]
      },
...
      "ReasonBreakdownDto": {
        "type": "object",
        "properties": {
          "regressionReasons": {
            "type": "object",
            "description": "Breakdown of regression reasons",
            "example": {
              "MISSING_GLOBAL_SIGNALS": 150,
              "BLOCKED_COUNTRY": 25,
              "NEUTRAL_LANGUAGE": 10
            },
            "additionalProperties": {
              "type": "number"
            }
          },
          "improvementReasons": {
            "type": "object",
            "description": "Breakdown of improvement reasons",
            "example": {
              "ALLOWED_COUNTRY": 100,
              "BREAKOUT_ALLOWED": 50
            },
            "additionalProperties": {
              "type": "number"
            }
          }
        },
        "required": [
          "regressionReasons",
          "improvementReasons"
        ]
      },
...
      "DryRunOptionsDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "description": "Selection mode for items to evaluate",
            "example": "sample",
            "enum": [
              "sample",
              "top",
              "byType",
              "byCountry"
            ]
          },
          "limit": {
            "type": "number",
            "description": "Maximum number of items to evaluate (default: 1000, max: 10000)",
            "example": 1000,
            "minimum": 1,
            "maximum": 10000
          },
          "mediaType": {
            "type": "string",
            "description": "Media type filter (required for byType mode)",
            "example": "movie",
            "enum": [
              "movie",
              "show"
            ]
          },
          "country": {
            "type": "string",
            "description": "Country filter (ISO 3166-1 alpha-2, required for byCountry mode)",
            "example": "US"
          },
          "samplePercent": {
            "type": "number",
            "description": "Sample percentage for sample mode (1-100)",
            "example": 10,
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "mode"
        ]
      },
      "DryRunRequestDto": {
        "type": "object",
        "properties": {
          "policy": {
            "description": "Proposed policy configuration to test",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreatePolicyDto"
              }
            ]
          },
          "options": {
            "description": "Dry-run options",
            "allOf": [
              {
                "$ref": "#/components/schemas/DryRunOptionsDto"
              }
            ]
          }
        },
        "required": [
          "policy",
          "options"
        ]
      },
      "DryRunSummaryDto": {
        "type": "object",
        "properties": {
          "totalEvaluated": {
            "type": "number",
            "description": "Total items evaluated",
            "example": 1000
          },
          "eligible": {
            "type": "number",
            "description": "Items that would be eligible",
            "example": 750
          },
          "ineligible": {
            "type": "number",
            "description": "Items that would be ineligible",
            "example": 200
          },
          "pending": {
            "type": "number",
            "description": "Items that would be pending",
            "example": 40
          },
          "review": {
            "type": "number",
            "description": "Items that would need review",
            "example": 10
          },
          "newlyEligible": {
            "type": "number",
            "description": "Items that would become newly eligible",
            "example": 50
          },
          "newlyIneligible": {
            "type": "number",
            "description": "Items that would become newly ineligible",
            "example": 25
          },
          "unchanged": {
            "type": "number",
            "description": "Items with unchanged status",
            "example": 925
          },
          "reasonBreakdown": {
            "description": "Breakdown of evaluation reasons",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReasonBreakdownDto"
            }
          },
          "executionTimeMs": {
            "type": "number",
            "description": "Execution time in milliseconds",
            "example": 1500
          },
          "mode": {
            "type": "string",
            "description": "Selection mode used",
            "example": "sample",
            "enum": [
              "sample",
              "top",
              "byType",
              "byCountry"
            ]
          },
          "limit": {
            "type": "number",
            "description": "Item limit used",
            "example": 1000
          }
        },
        "required": [
          "totalEvaluated",
          "eligible",
          "ineligible",
          "pending",
          "review",
          "newlyEligible",
          "newlyIneligible",
          "unchanged",
          "reasonBreakdown",
          "executionTimeMs",
          "mode",
          "limit"
        ]
      },
      "DryRunItemResultDto": {
        "type": "object",
        "properties": {
          "mediaItemId": {
            "type": "string",
            "description": "Media item ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Media item title",
            "example": "The Matrix"
          },
          "currentStatus": {
            "type": "string",
            "description": "Current eligibility status (null if not evaluated)",
            "example": "ELIGIBLE",
            "enum": [
              "PENDING",
              "ELIGIBLE",
              "INELIGIBLE",
              "REVIEW"
            ]
          },
          "proposedStatus": {
            "type": "string",
            "description": "Proposed eligibility status under new policy",
            "example": "ELIGIBLE",
            "enum": [
              "PENDING",
              "ELIGIBLE",
              "INELIGIBLE",
              "REVIEW"
            ]
          },
          "reasons": {
            "description": "Evaluation reasons",
            "example": [
              "ALLOWED_COUNTRY",
              "ALLOWED_LANGUAGE"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "relevanceScore": {
            "type": "number",
            "description": "Relevance score (0-100)",
            "example": 75
          },
          "breakoutRuleId": {
            "type": "string",
            "description": "Breakout rule ID if applicable",
            "example": "GLOBAL_HIT"
          },
          "statusChanged": {
            "type": "boolean",
            "description": "Whether status changed from current",
            "example": false
          }
        },
        "required": [
          "mediaItemId",
          "title",
          "proposedStatus",
          "reasons",
          "relevanceScore",
          "statusChanged"
        ]
      },
      "DryRunResponseDto": {
        "type": "object",
        "properties": {
          "summary": {
            "description": "Dry-run summary",
            "allOf": [
              {
                "$ref": "#/components/schemas/DryRunSummaryDto"
              }
            ]
          },
          "items": {
            "description": "Individual item results",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DryRunItemResultDto"
            }
          },
          "currentPolicyVersion": {
            "type": "number",
            "description": "Current active policy version (for diff mode)",
            "example": 1
          }
        },
        "required": [
          "summary",
          "items"
        ]
      },
      "BackfillResponseDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}